(fix) Only include one *-ujs library
In Rails 5.1 `rails-ujs` replaced `jquery-ujs` to remove the dependency
on jQuery. It is a like for like replacement, this means by including
both of them in our project we're double up on the same event listeners.
This caused errors in Firefox and IE when the user clicked a click with
the `data-confirm` attribute, they would get two confirmation boxes -
then clicking the second one would result in being redirected to a 404
page.

This also means we can remove the manual accepting of the delete
confirmation which was a side-effect of having two confirms